Task_20180119_2
If a record has no parent record in database, insert its parent record
into database recursively until reaches Post level. For both Facebook
and Twitter

a.	If a comment has keyword & Geo but its parent (a post) does not in
db, then save it.
b.	If a reply has keyword & Geo but its parent (a comment) does not in
db, then save it (comment) and then call step a recursively to save
parent post.